Make sure changing repo starts in a clean state & implemented tab lazy loading
Changing the open repository in the current tab will now create a new tab that replaces the current one instead of updating the TabViewModel and having to make sure that every subviewmodel handles the change of repository properly.

This also allowed implementing lazy loading of a tab, so it does not load every repo at the same time.

Fixes #122