Visually distinguish log divider from commit-graph line.
The divider between the logs' commit graph and messages sections
used the the primary accent color, which is bright and powerful,
just like the commit-graph lines right next to it.

To create visual difference between the commit graph and the
draggable divider, make the divider less noticeable by
coloring it in the background color.
Also make it wider, so it becomes more like the padding elsewhere.
This makes it clearer that it's a UI thing, not a graph thing.